server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "fastapi[standard]>=0.134.0",
    "pydantic-settings>=2.13.1",
    "uvicorn>=0.41.0",
]
